@@ -1,9 +1,16 @@
 module golang.org/x/oauth2
 
-go 1.11
+go 1.18
 
 require (
-	cloud.google.com/go v0.65.0
-	golang.org/x/net v0.0.0-20220127200216-cd36cc0744dd
-	google.golang.org/appengine v1.6.6
+	cloud.google.com/go/compute/metadata v0.2.3
+	github.com/google/go-cmp v0.5.9
+	google.golang.org/appengine v1.6.7
+)
+
+require (
+	cloud.google.com/go/compute v1.20.1 // indirect
+	github.com/golang/protobuf v1.5.3 // indirect
+	golang.org/x/net v0.14.0 // indirect
+	google.golang.org/protobuf v1.31.0 // indirect
 )
